Refactor components into folders
